refactor(rpc): split rpc server module
