+++ bin/strata-bridge/src/mode/mod.rs
@@ -1,4 +1,5 @@
 //! The `mode` module contains the different modes of operation for the Strata Bridge.
 pub(crate) mod operator;
+pub(crate) mod rpc_server;
 pub(super) mod services;
 pub(crate) mod watchtower;
