+++ bin/strata-bridge/src/mode/rpc_server/mod.rs
@@ -1,0 +1,9 @@
+//! RPC server initialization and monitoring response helpers.
+
+mod monitoring;
+mod server;
+
+#[cfg(test)]
+mod tests;
+
+pub(in crate::mode) use server::init_rpc_server;
