+++ bin/strata-bridge/src/mode/rpc_server/tests/mod.rs
@@ -1,0 +1,3 @@
+//! RPC server module tests.
+
+mod monitoring;
