+++ bin/strata-bridge/src/mode/services/mod.rs
@@ -7,5 +7,4 @@ pub(in crate::mode) mod operator_table;
 pub(in crate::mode) mod operator_wallet;
 pub(in crate::mode) mod orchestrator;
 pub(in crate::mode) mod p2p_handles;
-pub(in crate::mode) mod rpc_server;
 pub(in crate::mode) mod secret_service;
